refactor: isolate file dialog provider
