+++ UserInterface/Widget/GameDropCard.py
@@ -12,7 +12,7 @@ from qfluentwidgets import BodyLabel, CardWidget, CaptionLabel, FlowLayout, Flue
 from ModuleFolders.Base.Base import Base
 from ModuleFolders.Config.Config import ConfigMixin
 from ModuleFolders.Infrastructure.Platform.PlatformPaths import monospace_font_family, ui_font_family
-from UserInterface.Native.MacOSUI import choose_input_folder_title, get_existing_directory, input_folder_button_text
+from UserInterface.Native.FileDialogProvider import get_existing_directory
 
 class InfoBlockWidget(ConfigMixin, Base, QWidget):
     """信息块组件 (圆角+折角+透明)"""
@@ -206,7 +206,7 @@ class DragDropArea(ConfigMixin, Base, QWidget):
         self.NoneLabel2 = CaptionLabel(f"       ", self)
         self.NoneLabel2.setAlignment(Qt.AlignVCenter) # 垂直居中        
 
-        info = input_folder_button_text(self.tra)
+        info = self.tra("拖拽/选择输入文件夹")
         self.selectButton = PrimaryPushButton(FluentIcon.FOLDER_ADD,info,self) # 创建主操作按钮
         self.selectButton.clicked.connect(self._select_folder) # 连接按钮点击事件到选择文件夹方法
 
@@ -251,8 +251,8 @@ class DragDropArea(ConfigMixin, Base, QWidget):
 
     def _select_folder(self):
         """处理点击选择文件夹按钮的事件"""
-        info = choose_input_folder_title(self.tra)
-        folder_path = get_existing_directory(self, info, self.current_path)
+        info = self.tra("选择文件夹")
+        folder_path = get_existing_directory(self, info)
         if folder_path:
             self.update_path(folder_path)
             button_center_global = self.selectButton.mapToGlobal(self.selectButton.rect().center())
